Adjust lint configs
