--- packages/.new-package/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "root": true,
-  "extends": ["../../.eslintrc-typescript.json"]
-}
