--- packages/api-client/__tests__/hello.test.ts
@@ -1,5 +1,0 @@
-import { it, expect } from '@jest/globals';
-
-it('should work', () => {
-  expect(true).not.toBe(false);
-});
