--- packages/api-client/__tests__/setup.ts
@@ -1,2 +1,0 @@
-// suppress warning about empty config
-process.env.SUPPRESS_NO_CONFIG_WARNING = 'true';
