--- packages/api-client/jest.config.js
@@ -1,3 +1,0 @@
-const { SWC } = require('../../common.jest.config');
-
-module.exports = SWC;
