+++ packages/api-client/package.json
@@ -12,22 +12,21 @@
   "author": "Beyond Essential Systems Pty. Ltd.",
   "license": "SEE LICENSE IN ../../license",
   "scripts": {
-    "build": "swc --delete-dir-on-start --out-dir dist --copy-files --source-maps true src",
-    "build-watch": "yarn run build --watch",
-    "test": "NODE_ENV=test jest",
-    "lint": "eslint src"
+    "build": "yarn run build:src && yarn run build:types",
+    "build:src": "swc --delete-dir-on-start --out-dir dist --copy-files --source-maps true src",
+    "build:types": "tsc --outDir dist --declaration --emitDeclarationOnly --noEmit false && mv dist/src/* dist && rm -rf dist/src",
+    "build-watch": "yarn run build:src --watch",
+    "lint": "yarn run lint:check && yarn run lint:types",
+    "lint:check": "eslint src",
+    "lint:types": "tsc"
   },
   "devDependencies": {
     "@beyondessential/eslint-config-beyondessential": "^2.6.0",
-    "@jest/globals": "^29.4.3",
-    "@swc/jest": "^0.2.24",
     "@tamanu/build-tooling": "*",
     "@types/node": "^18.14.6",
     "@typescript-eslint/eslint-plugin": "^6.2.0",
     "@typescript-eslint/parser": "^6.2.0",
-    "eslint": "^8.45.0",
-    "jest": "^29.4.3",
-    "typescript": "^5.1.6"
+    "eslint": "^8.45.0"
   },
   "dependencies": {
     "@tamanu/shared": "*",
