+++ packages/api-client/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": ["../../common.tsconfig.json"],
+  "include": [
+    "src/**/*.ts"
+  ],
+  "compilerOptions": {
+    "baseUrl": "./src",
+  }
+}
